requirements.txt
streamlit
pillow
pandas
numpy
opencv-contrib-python
opencv-python
openpyxl
pyttsx3

# haarcascade_frontalface_default.xml  ->>>>
# haarcascades - the folder contains trained classifiers for detecting objects
#             of a particular type, e.g. faces (frontal, profile), pedestrians etc.
#             Some of the classifiers have a special license - please,
#             look into the files for details.
# https://github.com/opencv/opencv/tree/master/data/haarcascades

